+++ src/Store/IStore.cs
@@ -1,0 +1,4 @@
+namespace AReSSO.Store
+{
+    public interface IStore<out TRootState> : IActionDispatcher, IStateContainer<TRootState> { }
+}