+++ usageExamples/ExamplePerson.cs
@@ -27,25 +27,9 @@ namespace AReSSO.UsageExamples
             Birthday = birthday;
             FavoriteSongs = favoriteSongs;
         }
-
-        /// <summary>
-        /// This is a private copy constructor. It is used to create a copy when Copy is called.
-        /// </summary>
-        /// <remarks>
-        /// This method is basically entirely boilerplate and could easily be automated.
-        /// </remarks>
-        private ExamplePerson(ExamplePerson old,
-            PropertyChange<string> newName,
-            PropertyChange<DateTime> newBirthday,
-            PropertyChange<IReadOnlyList<ExampleSong>> newFavoriteSongs)
-        {
-            Name = newName.Else(old.Name);
-            Birthday = newBirthday.Else(old.Birthday);
-            FavoriteSongs = newFavoriteSongs.Else(old.FavoriteSongs);
-        }
         
         /// <summary>
-        /// The copy function is a pure function that returns a copy of a state.
+        /// The copy function is a pure function that returns a shallow copy of an object.
         /// Note the usage of optional arguments. PropertyDelta defaults to Changed = false, so a property will only
         /// change for properties that are specified.
         ///
@@ -59,8 +43,10 @@ namespace AReSSO.UsageExamples
         public ExamplePerson Copy(
             PropertyChange<string> name = default,
             PropertyChange<DateTime> birthday = default,
-            PropertyChange<IReadOnlyList<ExampleSong>> favoriteSongs = default) =>
-            new ExamplePerson(this, name, birthday, favoriteSongs);
+            PropertyChange<IReadOnlyList<ExampleSong>> favoriteSongs = default)
+        {
+            return new ExamplePerson(name.Else(Name), birthday.Else(Birthday), favoriteSongs.Else(FavoriteSongs));
+        }
 
         #region Equals Implementations
         // Generic implementations of Equals and GetHashCode.